# BackEnd/OnlineJudge/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/sort": {
            "post": {
                "description": "获取所有用户并按分数降序排序",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "根据用户分数降序排序",
                "parameters": [
                    {
                        "type": "string",
                        "description": "执行操作的用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取并排序用户列表",
                        "schema": {
                            "$ref": "#/definitions/models._ResponseSort"
                        }
                    },
                    "403": {
                        "description": "Token 已超时",
                        "schema": {
                            "$ref": "#/definitions/models._ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "数据库查询错误",
                        "schema": {
                            "$ref": "#/definitions/models._ResponseError"
                        }
                    }
                }
            }
        },
        "/changeScore": {
            "post": {
                "description": "通过用户名和新分数的表单数据，增加用户的总分数，并返回新的总分数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "增加用户分数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "要增加的分数",
                        "name": "newscore",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "增加分数成功，返回消息和新的总分数",
                        "schema": {
                            "$ref": "#/definitions/models._ResponseAddScore"
                        }
                    },
                    "403": {
                        "description": "Token 已超时",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "数据库查询或保存出错",
                        "schema": {
                            "$ref": "#/definitions/models._ResponseError"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "用户使用用户名和密码进行登录，成功返回Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功登录，返回Token",
                        "schema": {
                            "$ref": "#/definitions/models._ResponseToken"
                        }
                    },
                    "403": {
                        "description": "用户名不存在或密码错误",
                        "schema": {
                            "$ref": "#/definitions/models._ResponseMsg"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "用户使用用户名、密码和昵称进行注册，成功返回Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "昵称",
                        "name": "nickname",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功注册，返回Token和Token过期时间",
                        "schema": {
                            "$ref": "#/definitions/models._ResponseReg"
                        }
                    },
                    "403": {
                        "description": "用户已存在或注册失败",
                        "schema": {
                            "$ref": "#/definitions/models._ResponseRegErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models._ResponseAddScore": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "new_score": {
                    "type": "integer"
                }
            }
        },
...
        "models._ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models._ResponseMsg": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "models._ResponseReg": {
            "type": "object",
            "properties": {
                "Token": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                }
            }
        },
        "models._ResponseRegErr": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "models._ResponseSort": {
            "type": "object"
        },
        "models._ResponseToken": {
            "type": "object",
            "properties": {
                "Token": {
                    "type": "string"
                }
            }
        },
...
    }
}